First complete (working) and deployed version of the app
The Results component uses the `query-string` module to parse the url search string
and obtains the names of the two players to be compared. A Github Client ID and
Client Secret are appended to the GET request to fetch the player's profile details.
The webpack configuration is modified to be able to build for production. A script is
added to package.json to configure firebase credentials local to the project.

  - Add the Results component to compute the results of the battle and display
    the same

  - Add a highly reusable Loading component that can be displayed when a page
    is loading data

  - Modify webpack configuration (webpack.config.js) to facilitate building for
    production

  - Add firebase-tools to deploy to Firebase

  - Deploy to Firebase